skeleton/requirements/development.txt
web.py==0.36
wsgiref==0.1.2
redis==2.10.3
git+git://github.com/kiddouk/redisco.git#egg=redisco    # redisco
